Separated descriptors and fields from models; Added first 1.9 compatibility hacks; Most tests passing again
