+++ versions/models.py
@@ -20,25 +20,26 @@ import re
 
 from django import VERSION
 
+from versions.util import get_utc_now
+
 if VERSION[:2] >= (1, 8):
     from django.db.models.sql.datastructures import Join
 if VERSION[:2] >= (1, 7):
     from django.apps.registry import apps
 from django.core.exceptions import SuspiciousOperation, ObjectDoesNotExist
 from django.db import transaction
-from django.db.models.base import Model
 from django.db.models import Q
 from django.db.models.constants import LOOKUP_SEP
-from django.db.models.fields import FieldDoesNotExist
-from django.db.models.fields.related import (ForeignKey, ReverseSingleRelatedObjectDescriptor,
-                                             ReverseManyRelatedObjectsDescriptor, ManyToManyField,
-                                             ManyRelatedObjectsDescriptor, create_many_related_manager,
-                                             ForeignRelatedObjectsDescriptor, RECURSIVE_RELATIONSHIP_CONSTANT)
-from django.db.models.query import QuerySet, ValuesListQuerySet, ValuesQuerySet
+from django.db.models.fields.related import ForeignKey, ManyToManyField, RECURSIVE_RELATIONSHIP_CONSTANT
+
+from django.db.models.query import QuerySet
+if VERSION[:2] >= (1, 9):
+    from django.db.models.query import ModelIterable
+else:
+    from django.db.models.query import ValuesListQuerySet, ValuesQuerySet
 from django.db.models.signals import post_init
 from django.db.models.sql import Query
 from django.db.models.sql.where import ExtraWhere, WhereNode
-from django.utils.functional import cached_property
 from django.utils.timezone import utc
 from django.utils import six
 
@@ -46,10 +47,6 @@ from django.db import models, router
 
 from versions import settings as versions_settings
 from versions.exceptions import DeletionOfNonCurrentVersionError
-
-
-def get_utc_now():
-    return datetime.datetime.utcnow().replace(tzinfo=utc)
 
 
 QueryTime = namedtuple('QueryTime', 'time active')
@@ -285,153 +282,7 @@ class VersionManager(models.Manager):
         return self.uuid_valid_form_regex.match(uuid_string) is not None
 
 
-class VersionedWhereNode(WhereNode):
-    def as_sql(self, qn, connection):
-        """
-        This method identifies joined table aliases in order for VersionedExtraWhere.as_sql()
-        to be able to add time restrictions for those tables based on the VersionedQuery's
-        querytime value.
 
-        :param qn: In Django 1.7 & 1.8 this is a compiler; in 1.6, it's an instance-method
-        :param connection: A DB connection
-        :return: A tuple consisting of (sql_string, result_params)
-        """
-        # self.children is an array of VersionedExtraWhere-objects
-        for child in self.children:
-            if isinstance(child, VersionedExtraWhere) and not child.params:
-                try:
-                    # Django 1.7 & 1.8 handles compilers as objects
-                    _query = qn.query
-                except AttributeError:
-                    # Django 1.6 handles compilers as instancemethods
-                    _query = qn.__self__.query
-                query_time = _query.querytime.time
-                apply_query_time = _query.querytime.active
-                alias_map = _query.alias_map
-                # In Django 1.6 & 1.7, use the join_map to know, what *table* gets joined to which
-                # *left-hand sided* table
-                # In Django 1.8, use the Join objects in alias_map
-                if hasattr(_query, 'join_map'):
-                    self._set_child_joined_alias_using_join_map(child, _query.join_map, alias_map)
-                else:
-                    self._set_child_joined_alias(child, alias_map)
-                if apply_query_time:
-                    # Add query parameters that have not been added till now
-                    child.set_as_of(query_time)
-                else:
-                    # Remove the restriction if it's not required
-                    child.sqls = []
-        return super(VersionedWhereNode, self).as_sql(qn, connection)
-
-    @staticmethod
-    def _set_child_joined_alias_using_join_map(child, join_map, alias_map):
-        """
-        Set the joined alias on the child, for Django <= 1.7.x.
-        :param child:
-        :param join_map:
-        :param alias_map:
-        """
-        for lhs, table, join_cols in join_map:
-            if lhs is None:
-                continue
-            if lhs == child.alias:
-                relevant_alias = child.related_alias
-            elif lhs == child.related_alias:
-                relevant_alias = child.alias
-            else:
-                continue
-
-            join_info = alias_map[relevant_alias]
-            if join_info.join_type is None:
-                continue
-
-            if join_info.lhs_alias in [child.alias, child.related_alias]:
-                child.set_joined_alias(relevant_alias)
-                break
-
-    @staticmethod
-    def _set_child_joined_alias(child, alias_map):
-        """
-        Set the joined alias on the child, for Django >= 1.8.0
-        :param child:
-        :param alias_map:
-        """
-        for table in alias_map:
-            join = alias_map[table]
-            if not isinstance(join, Join):
-                continue
-            lhs = join.parent_alias
-            if (lhs == child.alias and table == child.related_alias) \
-                    or (lhs == child.related_alias and table == child.alias):
-                child.set_joined_alias(table)
-                break
-
-
-class VersionedExtraWhere(ExtraWhere):
-    """
-    A specific implementation of ExtraWhere;
-    Before as_sql can be called on an object, ensure that calls to
-    - set_as_of and
-    - set_joined_alias
-    have been done
-    """
-
-    def __init__(self, historic_sql, current_sql, alias, remote_alias):
-        super(VersionedExtraWhere, self).__init__(sqls=[], params=[])
-        self.historic_sql = historic_sql
-        self.current_sql = current_sql
-        self.alias = alias
-        self.related_alias = remote_alias
-        self._as_of_time_set = False
-        self.as_of_time = None
-        self._joined_alias = None
-
-    def set_as_of(self, as_of_time):
-        self.as_of_time = as_of_time
-        self._as_of_time_set = True
-
-    def set_joined_alias(self, joined_alias):
-        """
-        Takes the alias that is being joined to the query and applies the query
-        time constraint to its table
-
-        :param str joined_alias: The table name of the alias
-        """
-        self._joined_alias = joined_alias
-
-    def as_sql(self, qn=None, connection=None):
-        sql = ""
-        params = []
-
-        # Fail fast for inacceptable cases
-        if self._as_of_time_set and not self._joined_alias:
-            raise ValueError("joined_alias is not set, but as_of is; this is a conflict!")
-
-        # Set the SQL string in dependency of whether as_of_time was set or not
-        if self._as_of_time_set:
-            if self.as_of_time:
-                sql = self.historic_sql
-                params = [self.as_of_time] * 2
-                # 2 is the number of occurences of the timestamp in an as_of-filter expression
-            else:
-                # If as_of_time was set to None, we're dealing with a query for "current" values
-                sql = self.current_sql
-        else:
-            # No as_of_time has been set; Perhaps, as_of was not part of the query -> That's OK
-            pass
-
-        # By here, the sql string is defined if an as_of_time was provided
-        if self._joined_alias:
-            sql = sql.format(alias=self._joined_alias)
-
-        # Set the final sqls
-        # self.sqls needs to be set before the call to parent
-        if sql:
-            self.sqls = [sql]
-        else:
-            self.sqls = ["1=1"]
-        self.params = params
-        return super(VersionedExtraWhere, self).as_sql(qn, connection)
 
 
 class VersionedQuery(Query):
@@ -442,6 +293,7 @@ class VersionedQuery(Query):
     """
 
     def __init__(self, *args, **kwargs):
+        from .fields import VersionedWhereNode
         kwargs['where'] = VersionedWhereNode
         super(VersionedQuery, self).__init__(*args, **kwargs)
         self.querytime = QueryTime(time=None, active=False)
@@ -570,9 +422,15 @@ class VersionedQuerySet(QuerySet):
         """
         if self._result_cache is None:
             self._result_cache = list(self.iterator())
-            if not isinstance(self, ValuesListQuerySet):
-                for x in self._result_cache:
-                    self._set_item_querytime(x)
+            if VERSION[:2] >= (1, 9):
+                # TODO: Do we have to test for ValuesListIterable, ValuesIterable, and FlatValuesListIterable here?
+                if self._iterable_class == ModelIterable:
+                    for x in self._result_cache:
+                        self._set_item_querytime(x)
+            else:
+                if not isinstance(self, ValuesListQuerySet):
+                    for x in self._result_cache:
+                        self._set_item_querytime(x)
         if self._prefetch_related_lookups and not self._prefetch_done:
             self._prefetch_related_objects()
 
@@ -666,544 +524,10 @@ class VersionedQuerySet(QuerySet):
     delete.queryset_only = True
 
 
-class VersionedForeignKey(ForeignKey):
-    """
-    We need to replace the standard ForeignKey declaration in order to be able to introduce
-    the VersionedReverseSingleRelatedObjectDescriptor, which allows to go back in time...
-    We also want to allow keeping track of any as_of time so that joins can be restricted
-    based on that.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(VersionedForeignKey, self).__init__(*args, **kwargs)
-
-    def contribute_to_class(self, cls, name, virtual_only=False):
-        super(VersionedForeignKey, self).contribute_to_class(cls, name, virtual_only)
-        setattr(cls, self.name, VersionedReverseSingleRelatedObjectDescriptor(self))
-
-    def contribute_to_related_class(self, cls, related):
-        """
-        Override ForeignKey's methods, and replace the descriptor, if set by the parent's methods
-        """
-        # Internal FK's - i.e., those with a related name ending with '+' -
-        # and swapped models don't get a related descriptor.
-        super(VersionedForeignKey, self).contribute_to_related_class(cls, related)
-        accessor_name = related.get_accessor_name()
-        if hasattr(cls, accessor_name):
-            setattr(cls, accessor_name, VersionedForeignRelatedObjectsDescriptor(related))
-
-    def get_extra_restriction(self, where_class, alias, remote_alias):
-        """
-        Overrides ForeignObject's get_extra_restriction function that returns an SQL statement which is appended to a
-        JOIN's conditional filtering part
-
-        :return: SQL conditional statement
-        :rtype: WhereNode
-        """
-        historic_sql = '''{alias}.version_start_date <= %s
-                 AND ({alias}.version_end_date > %s OR {alias}.version_end_date is NULL )'''
-        current_sql = '''{alias}.version_end_date is NULL'''
-        # How 'bout creating an ExtraWhere here, without params
-        return where_class([VersionedExtraWhere(historic_sql=historic_sql, current_sql=current_sql, alias=alias,
-                                                remote_alias=remote_alias)])
-
-    def get_joining_columns(self, reverse_join=False):
-        """
-        Get and return joining columns defined by this foreign key relationship
-
-        :return: A tuple containing the column names of the tables to be joined (<local_col_name>, <remote_col_name>)
-        :rtype: tuple
-        """
-        source = self.reverse_related_fields if reverse_join else self.related_fields
-        joining_columns = tuple()
-        for lhs_field, rhs_field in source:
-            lhs_col_name = lhs_field.column
-            rhs_col_name = rhs_field.column
-            # Test whether
-            # - self is the current ForeignKey relationship
-            # - self was not auto_created (e.g. is not part of a M2M relationship)
-            if self is lhs_field and not self.auto_created:
-                if rhs_col_name == Versionable.VERSION_IDENTIFIER_FIELD:
-                    rhs_col_name = Versionable.OBJECT_IDENTIFIER_FIELD
-            elif self is rhs_field and not self.auto_created:
-                if lhs_col_name == Versionable.VERSION_IDENTIFIER_FIELD:
-                    lhs_col_name = Versionable.OBJECT_IDENTIFIER_FIELD
-            joining_columns = joining_columns + ((lhs_col_name, rhs_col_name),)
-        return joining_columns
-
-
-class VersionedManyToManyField(ManyToManyField):
-    def __init__(self, *args, **kwargs):
-        super(VersionedManyToManyField, self).__init__(*args, **kwargs)
-
-    def contribute_to_class(self, cls, name):
-        """
-        Called at class type creation. So, this method is called, when metaclasses get created
-        """
-        # self.rel.through needs to be set prior to calling super, since super(...).contribute_to_class refers to it.
-        # Classes pointed to by a string do not need to be resolved here, since Django does that at a later point in
-        # time - which is nice... ;)
-        #
-        # Superclasses take care of:
-        # - creating the through class if unset
-        # - resolving the through class if it's a string
-        # - resolving string references within the through class
-        if not self.rel.through and not cls._meta.abstract and not cls._meta.swapped:
-            self.rel.through = VersionedManyToManyField.create_versioned_many_to_many_intermediary_model(self, cls,
-                                                                                                         name)
-        super(VersionedManyToManyField, self).contribute_to_class(cls, name)
-
-        # Overwrite the descriptor
-        if hasattr(cls, self.name):
-            setattr(cls, self.name, VersionedReverseManyRelatedObjectsDescriptor(self))
-
-    def contribute_to_related_class(self, cls, related):
-        """
-        Called at class type creation. So, this method is called, when metaclasses get created
-        """
-        super(VersionedManyToManyField, self).contribute_to_related_class(cls, related)
-        accessor_name = related.get_accessor_name()
-        if accessor_name and hasattr(cls, accessor_name):
-            descriptor = VersionedManyRelatedObjectsDescriptor(related, accessor_name)
-            setattr(cls, accessor_name, descriptor)
-            if hasattr(cls._meta, 'many_to_many_related') and isinstance(cls._meta.many_to_many_related, list):
-                cls._meta.many_to_many_related.append(descriptor)
-            else:
-                cls._meta.many_to_many_related = [descriptor]
-
-    @staticmethod
-    def create_versioned_many_to_many_intermediary_model(field, cls, field_name):
-        # Let's not care too much on what flags could potentially be set on that intermediary class (e.g. managed, etc)
-        # Let's play the game, as if the programmer had specified a class within his models... Here's how.
-
-        from_ = cls._meta.model_name
-        to_model = field.rel.to
-
-        # Force 'to' to be a string (and leave the hard work to Django)
-        if not isinstance(field.rel.to, six.string_types):
-            to_model = '%s.%s' % (field.rel.to._meta.app_label, field.rel.to._meta.object_name)
-            to = field.rel.to._meta.object_name.lower()
-        else:
-            to = to_model.lower()
-        name = '%s_%s' % (from_, field_name)
-
-        if field.rel.to == RECURSIVE_RELATIONSHIP_CONSTANT or to == cls._meta.object_name:
-            from_ = 'from_%s' % to
-            to = 'to_%s' % to
-            to_model = cls
-
-        # Since Django 1.7, a migration mechanism is shipped by default with Django. This migration module loads all
-        # declared apps' models inside a __fake__ module.
-        # This means that the models can be already loaded and registered by their original module, when we
-        # reach this point of the application and therefore there is no need to load them a second time.
-        if VERSION[:2] >= (1, 7) and cls.__module__ == '__fake__':
-            try:
-                # Check the apps for an already registered model
-                return apps.get_registered_model(cls._meta.app_label, str(name))
-            except KeyError:
-                # The model has not been registered yet, so continue
-                pass
-
-        meta = type('Meta', (object,), {
-            # 'unique_together': (from_, to),
-            'auto_created': cls,
-            'db_tablespace': cls._meta.db_tablespace,
-            'app_label': cls._meta.app_label,
-        })
-        return type(str(name), (Versionable,), {
-            'Meta': meta,
-            '__module__': cls.__module__,
-            from_: VersionedForeignKey(cls, related_name='%s+' % name, auto_created=name),
-            to: VersionedForeignKey(to_model, related_name='%s+' % name, auto_created=name),
-        })
-
-
-class VersionedReverseSingleRelatedObjectDescriptor(ReverseSingleRelatedObjectDescriptor):
-    """
-    A ReverseSingleRelatedObjectDescriptor-typed object gets inserted, when a ForeignKey
-    is defined in a Django model. This is one part of the analogue for versioned items.
-
-    Unfortunately, we need to run two queries. The first query satisfies the foreign key
-    constraint. After extracting the identity information and combining it with the datetime-
-    stamp, we are able to fetch the historic element.
-    """
-
-    def __get__(self, instance, instance_type=None):
-        """
-        The getter method returns the object, which points instance, e.g. choice.poll returns
-        a Poll instance, whereas the Poll class defines the ForeignKey.
-        :param instance: The object on which the property was accessed
-        :param instance_type: The type of the instance object
-        :return: Returns a Versionable
-        """
-        current_elt = super(VersionedReverseSingleRelatedObjectDescriptor, self).__get__(instance, instance_type)
-
-        if instance is None:
-            return self
-
-        if not current_elt:
-            return None
-
-        if not isinstance(current_elt, Versionable):
-            raise TypeError("VersionedForeignKey target is of type "
-                            + str(type(current_elt))
-                            + ", which is not a subclass of Versionable")
-
-        if hasattr(instance, '_querytime'):
-            # If current_elt matches the instance's querytime, there's no need to make a database query.
-            if Versionable.matches_querytime(current_elt, instance._querytime):
-                current_elt._querytime = instance._querytime
-                return current_elt
-
-            return current_elt.__class__.objects.as_of(instance._querytime.time).get(identity=current_elt.identity)
-        else:
-            return current_elt.__class__.objects.current.get(identity=current_elt.identity)
-
-
-class VersionedForeignRelatedObjectsDescriptor(ForeignRelatedObjectsDescriptor):
-    """
-    This descriptor generates the manager class that is used on the related object of a ForeignKey relation
-    (i.e. the reverse-ForeignKey field manager).
-    """
-
-    @cached_property
-    def related_manager_cls(self):
-        # return create_versioned_related_manager
-        manager_cls = super(VersionedForeignRelatedObjectsDescriptor, self).related_manager_cls
-        rel_field = self.related.field
-
-        class VersionedRelatedManager(manager_cls):
-            def __init__(self, instance):
-                super(VersionedRelatedManager, self).__init__(instance)
-
-                # This is a hack, in order to get the versioned related objects
-                for key in self.core_filters.keys():
-                    if '__exact' in key or '__' not in key:
-                        self.core_filters[key] = instance.identity
-
-            def get_queryset(self):
-                queryset = super(VersionedRelatedManager, self).get_queryset()
-                # Do not set the query time if it is already correctly set.  queryset.as_of() returns a clone
-                # of the queryset, and this will destroy the prefetched objects cache if it exists.
-                if isinstance(queryset,
-                              VersionedQuerySet) and self.instance._querytime.active and queryset.querytime != self.instance._querytime:
-                    queryset = queryset.as_of(self.instance._querytime.time)
-                return queryset
-
-            def add(self, *objs):
-                cloned_objs = ()
-                for obj in objs:
-                    if not isinstance(obj, Versionable):
-                        raise TypeError("Trying to add a non-Versionable to a VersionedForeignKey relationship")
-                    cloned_objs += (obj.clone(),)
-                super(VersionedRelatedManager, self).add(*cloned_objs)
-
-            # clear() and remove() are present if the FK is nullable
-            if 'clear' in dir(manager_cls):
-                def clear(self, **kwargs):
-                    """
-                    Overridden to ensure that the current queryset is used, and to clone objects before they
-                    are removed, so that history is not lost.
-                    """
-                    bulk = kwargs.pop('bulk', True)
-                    db = router.db_for_write(self.model, instance=self.instance)
-                    queryset = self.current.using(db)
-                    with transaction.atomic(using=db, savepoint=False):
-                        cloned_pks = [obj.clone().pk for obj in queryset]
-                        update_qs = self.current.filter(pk__in=cloned_pks)
-                        if VERSION[:2] == (1, 6):
-                            update_qs.update(**{rel_field.name: None})
-                        else:
-                            self._clear(update_qs, bulk)
-
-            if 'remove' in dir(manager_cls):
-                def remove(self, *objs):
-                    val = rel_field.get_foreign_related_value(self.instance)
-                    cloned_objs = ()
-                    for obj in objs:
-                        # Is obj actually part of this descriptor set? Otherwise, silently go over it, since Django
-                        # handles that case
-                        if rel_field.get_local_related_value(obj) == val:
-                            # Silently pass over non-versionable items
-                            if not isinstance(obj, Versionable):
-                                raise TypeError(
-                                    "Trying to remove a non-Versionable from a VersionedForeignKey realtionship")
-                            cloned_objs += (obj.clone(),)
-                    super(VersionedRelatedManager, self).remove(*cloned_objs)
-
-        return VersionedRelatedManager
-
-
-def create_versioned_many_related_manager(superclass, rel):
-    """
-    The "casting" which is done in this method is needed, since otherwise, the methods introduced by
-    Versionable are not taken into account.
-    :param superclass: This is usually a models.Manager
-    :param rel: Contains the ManyToMany relation
-    :return: A subclass of ManyRelatedManager and Versionable
-    """
-    many_related_manager_klass = create_many_related_manager(superclass, rel)
-
-    class VersionedManyRelatedManager(many_related_manager_klass):
-        def __init__(self, *args, **kwargs):
-            super(VersionedManyRelatedManager, self).__init__(*args, **kwargs)
-            # Additional core filters are: version_start_date <= t & (version_end_date > t | version_end_date IS NULL)
-            # but we cannot work with the Django core filters, since they don't support ORing filters, which
-            # is a thing we need to consider the "version_end_date IS NULL" case;
-            # So, we define our own set of core filters being applied when versioning
-            try:
-                version_start_date_field = self.through._meta.get_field('version_start_date')
-                version_end_date_field = self.through._meta.get_field('version_end_date')
-            except FieldDoesNotExist as e:
-                if VERSION[:2] >= (1, 8):
-                    fields = [f.name for f in self.through._meta.get_fields()]
-                else:
-                    fields = self.through._meta.get_all_field_names()
-                print(str(e) + "; available fields are " + ", ".join(fields))
-                raise e
-                # FIXME: this probably does not work when auto-referencing
-
-        def get_queryset(self):
-            """
-            Add a filter to the queryset, limiting the results to be pointed by relationship that are
-            valid for the given timestamp (which is taken at the current instance, or set to now, if not
-            available).
-            Long story short, apply the temporal validity filter also to the intermediary model.
-            """
-
-            queryset = super(VersionedManyRelatedManager, self).get_queryset()
-            if hasattr(queryset, 'querytime'):
-                if self.instance._querytime.active and self.instance._querytime != queryset.querytime:
-                    queryset = queryset.as_of(self.instance._querytime.time)
-            return queryset
-
-        def _remove_items(self, source_field_name, target_field_name, *objs):
-            """
-            Instead of removing items, we simply set the version_end_date of the current item to the
-            current timestamp --> t[now].
-            Like that, there is no more current entry having that identity - which is equal to
-            not existing for timestamps greater than t[now].
-            """
-            return self._remove_items_at(None, source_field_name, target_field_name, *objs)
-
-        def _remove_items_at(self, timestamp, source_field_name, target_field_name, *objs):
-            if objs:
-                if timestamp is None:
-                    timestamp = get_utc_now()
-                old_ids = set()
-                for obj in objs:
-                    if isinstance(obj, self.model):
-                        # The Django 1.7-way is preferred
-                        if hasattr(self, 'target_field'):
-                            fk_val = self.target_field.get_foreign_related_value(obj)[0]
-                        # But the Django 1.6.x -way is supported for backward compatibility
-                        elif hasattr(self, '_get_fk_val'):
-                            fk_val = self._get_fk_val(obj, target_field_name)
-                        else:
-                            raise TypeError("We couldn't find the value of the foreign key, this might be due to the "
-                                            "use of an unsupported version of Django")
-                        old_ids.add(fk_val)
-                    else:
-                        old_ids.add(obj)
-                db = router.db_for_write(self.through, instance=self.instance)
-                qs = self.through._default_manager.using(db).filter(**{
-                    source_field_name: self.instance.id,
-                    '%s__in' % target_field_name: old_ids
-                }).as_of(timestamp)
-                for relation in qs:
-                    relation._delete_at(timestamp)
-
-        if 'add' in dir(many_related_manager_klass):
-            def add(self, *objs):
-                if not self.instance.is_current:
-                    raise SuspiciousOperation(
-                        "Adding many-to-many related objects is only possible on the current version")
-
-                # The ManyRelatedManager.add() method uses the through model's default manager to get
-                # a queryset when looking at which objects already exist in the database.
-                # In order to restrict the query to the current versions when that is done,
-                # we temporarily replace the queryset's using method so that the version validity
-                # condition can be specified.
-                klass = self.through._default_manager.get_queryset().__class__
-                __using_backup = klass.using
-
-                def using_replacement(self, *args, **kwargs):
-                    qs = __using_backup(self, *args, **kwargs)
-                    return qs.as_of(None)
-
-                klass.using = using_replacement
-                super(VersionedManyRelatedManager, self).add(*objs)
-                klass.using = __using_backup
-
-            def add_at(self, timestamp, *objs):
-                """
-                This function adds an object at a certain point in time (timestamp)
-                """
-                # First off, define the new constructor
-                def _through_init(self, *args, **kwargs):
-                    super(self.__class__, self).__init__(*args, **kwargs)
-                    self.version_birth_date = timestamp
-                    self.version_start_date = timestamp
-
-                # Through-classes have an empty constructor, so it can easily be overwritten when needed;
-                # This is not the default case, so the overwrite only takes place when we "modify the past"
-                self.through.__init_backup__ = self.through.__init__
-                self.through.__init__ = _through_init
-
-                # Do the add operation
-                self.add(*objs)
-
-                # Remove the constructor again (by replacing it with the original empty constructor)
-                self.through.__init__ = self.through.__init_backup__
-                del self.through.__init_backup__
-
-            add_at.alters_data = True
-
-        if 'remove' in dir(many_related_manager_klass):
-            def remove_at(self, timestamp, *objs):
-                """
-                Performs the act of removing specified relationships at a specified time (timestamp);
-                So, not the objects at a given time are removed, but their relationship!
-                """
-                self._remove_items_at(timestamp, self.source_field_name, self.target_field_name, *objs)
-
-                # For consistency, also handle the symmetrical case
-                if self.symmetrical:
-                    self._remove_items_at(timestamp, self.target_field_name, self.source_field_name, *objs)
-
-            remove_at.alters_data = True
-
-    return VersionedManyRelatedManager
-
-
-class VersionedReverseManyRelatedObjectsDescriptor(ReverseManyRelatedObjectsDescriptor):
-    """
-    Beside having a very long name, this class is useful when it comes to versioning the
-    ReverseManyRelatedObjectsDescriptor (huhu!!). The main part is the exposure of the
-    'related_manager_cls' property
-    """
-
-    def __get__(self, instance, owner=None):
-        """
-        Reads the property as which this object is figuring; mainly used for debugging purposes
-        :param instance: The instance on which the getter was called
-        :param owner: no idea... alternatively called 'instance_type by the superclasses
-        :return: A VersionedManyRelatedManager object
-        """
-        return super(VersionedReverseManyRelatedObjectsDescriptor, self).__get__(instance, owner)
-
-    def __set__(self, instance, value):
-        """
-        Completely overridden to avoid bulk deletion that happens when the parent method calls clear().
-
-        The parent method's logic is basically: clear all in bulk, then add the given objects in bulk.
-        Instead, we figure out which ones are being added and removed, and call add and remove for these values.
-        This lets us retain the versioning information.
-
-        Since this is a many-to-many relationship, it is assumed here that the django.db.models.deletion.Collector
-        logic, that is used in clear(), is not necessary here.  Collector collects related models, e.g. ones that should
-        also be deleted because they have a ON CASCADE DELETE relationship to the object, or, in the case of
-        "Multi-table inheritance", are parent objects.
-
-        :param instance: The instance on which the getter was called
-        :param value: iterable of items to set
-        """
-
-        if not instance.is_current:
-            raise SuspiciousOperation(
-                "Related values can only be directly set on the current version of an object")
-
-        if not self.field.rel.through._meta.auto_created:
-            opts = self.field.rel.through._meta
-            raise AttributeError(("Cannot set values on a ManyToManyField which specifies an intermediary model. "
-                                  "Use %s.%s's Manager instead.") % (opts.app_label, opts.object_name))
-
-        manager = self.__get__(instance)
-        # Below comment is from parent __set__ method.  We'll force evaluation, too:
-        # clear() can change expected output of 'value' queryset, we force evaluation
-        # of queryset before clear; ticket #19816
-        value = tuple(value)
-
-        being_removed, being_added = self.get_current_m2m_diff(instance, value)
-        timestamp = get_utc_now()
-        manager.remove_at(timestamp, *being_removed)
-        manager.add_at(timestamp, *being_added)
-
-    def get_current_m2m_diff(self, instance, new_objects):
-        """
-        :param instance: Versionable object
-        :param new_objects: objects which are about to be associated with instance
-        :return: (being_removed id list, being_added id list)
-        :rtype : tuple
-        """
-        new_ids = self.pks_from_objects(new_objects)
-        relation_manager = self.__get__(instance)
-
-        filter = Q(**{relation_manager.source_field.attname: instance.pk})
-        qs = self.through.objects.current.filter(filter)
-        try:
-            # Django 1.7
-            target_name = relation_manager.target_field.attname
-        except AttributeError:
-            # Django 1.6
-            target_name = relation_manager.through._meta.get_field_by_name(
-                relation_manager.target_field_name)[0].attname
-        current_ids = set(qs.values_list(target_name, flat=True))
-
-        being_removed = current_ids - new_ids
-        being_added = new_ids - current_ids
-        return list(being_removed), list(being_added)
-
-    def pks_from_objects(self, objects):
-        """
-        Extract all the primary key strings from the given objects.  Objects may be Versionables, or bare primary keys.
-        :rtype : set
-        """
-        return {o.pk if isinstance(o, Model) else o for o in objects}
-
-    @cached_property
-    def related_manager_cls(self):
-        return create_versioned_many_related_manager(
-            self.field.rel.to._default_manager.__class__,
-            self.field.rel
-        )
-
-
-class VersionedManyRelatedObjectsDescriptor(ManyRelatedObjectsDescriptor):
-    """
-    Beside having a very long name, this class is useful when it comes to versioning the
-    ManyRelatedObjectsDescriptor (huhu!!). The main part is the exposure of the
-    'related_manager_cls' property
-    """
-
-    via_field_name = None
-
-    def __init__(self, related, via_field_name):
-        super(VersionedManyRelatedObjectsDescriptor, self).__init__(related)
-        self.via_field_name = via_field_name
-
-    def __get__(self, instance, owner=None):
-        """
-        Reads the property as which this object is figuring; mainly used for debugging purposes
-        :param instance: The instance on which the getter was called
-        :param owner: no idea... alternatively called 'instance_type by the superclasses
-        :return: A VersionedManyRelatedManager object
-        """
-        return super(VersionedManyRelatedObjectsDescriptor, self).__get__(instance, owner)
-
-    @cached_property
-    def related_manager_cls(self):
-        return create_versioned_many_related_manager(
-            self.related.model._default_manager.__class__,
-            self.related.field.rel
-        )
-
-
 class Versionable(models.Model):
     """
     This is pretty much the central point for versioning objects.
     """
-
     VERSION_IDENTIFIER_FIELD = 'id'
     OBJECT_IDENTIFIER_FIELD = 'identity'
     VERSIONABLE_FIELDS = [VERSION_IDENTIFIER_FIELD, OBJECT_IDENTIFIER_FIELD, 'version_start_date',
@@ -1495,7 +819,10 @@ class Versionable(models.Model):
         opts = self._meta
         rel_field_names = [field.attname for field in opts.many_to_many]
         if hasattr(opts, 'many_to_many_related'):
-            rel_field_names += [rel.via_field_name for rel in opts.many_to_many_related]
+            if VERSION[:2] >= (1, 9):
+                rel_field_names += [rel.reverse for rel in opts.many_to_many_related]
+            else:
+                rel_field_names += [rel.via_field_name for rel in opts.many_to_many_related]
 
         return rel_field_names
 
@@ -1515,23 +842,6 @@ class Versionable(models.Model):
         self.version_start_date = self.version_birth_date = get_utc_now()
         self.version_end_date = None
         return self
-
-    @staticmethod
-    def matches_querytime(instance, querytime):
-        """
-        Checks whether the given instance satisfies the given QueryTime object.
-
-        :param instance: an instance of Versionable
-        :param querytime: QueryTime value to check against
-        """
-        if not querytime.active:
-            return True
-
-        if not querytime.time:
-            return instance.version_end_date is None
-
-        return (instance.version_start_date <= querytime.time
-                and (instance.version_end_date is None or instance.version_end_date > querytime.time))
 
 
 class VersionedManyToManyModel(object):
@@ -1562,3 +872,5 @@ class VersionedManyToManyModel(object):
 
 
 post_init.connect(VersionedManyToManyModel.post_init_initialize)
+
+
