+++ versions/util/__init__.py
@@ -1,0 +1,7 @@
+import datetime
+
+from django.utils.timezone import utc
+
+
+def get_utc_now():
+    return datetime.datetime.utcnow().replace(tzinfo=utc)
